Restructuring and Responsiveness
With some feedback started the restructure of the home page and creation of an about me page. Work in progress but for now it's a start.

Also fixed responsiveness and look of both the about me page and sources page. Restructured the nav bar and footer links

New project was added to the projects age, Citadel Level Design.

CSS changes and moving code to appropriate section.