Project Functionality Added, QOL Changes
Now when you click Doc on the home page for a project instead of just loading the default project on the project page you get the correct project!

Projects with no viewable source no longer have the view button shown as an option to click.

Also changed colors for tabs in the game html and CSS fix for the about me page with the floating footer.