Front Page Change, CSS Changes
CSS overhaul to accompany the restructuring of the front page. Added new headers and content to give details on the projects. Replaced Doc with Detail for clearer message as to what button does.

Hid, JS no longer being used by the front page, may find use for it with sliding pics/vid for each project.