Slideshow Update, and Code Cleanup
Updated the slideshow with more arguments and changed the logic a bit so now it can work with other projects.

Portfolio section now has images in the Screenshot section.

Went through some pages and got rid of unneeded console.logs, redundant code or styling/scripts. And removed an unused image.